Add new playbooks and change return value nfsServer.py
Add playbooks showing work with the terminal and group_vars parameters.
Divide all playbooks into create and delete parts.
Add to nfsServer_create output that only shows id of the created nfsServer.
Add a new function to the module nfsServer.py that adds
to the return value of the playbook id of the created nfsServer